[refactoring] TestManager Singleton 적용, 빌드 시간 감소를 위한 Testcase 분리
