+++ TeamB_Shell/ShellTest.h
@@ -18,7 +18,6 @@ class IShell {
 class ShellTest : public IShell {
  public:
   ShellTest() {}
-  ShellTest(TestManager *tm) : testManager(tm) {}
   void executeCommand(const std::string &cmdLine);
   std::string getOutput();
   std::string readFromFile(const std::string &file);
@@ -38,6 +37,4 @@ class ShellTest : public IShell {
   bool checkValidLba(std::string &lbaStr);
   bool checkValidValue(std::string &valueStr);
   bool checkValidSize(std::string &valueStr);
-
-  TestManager *testManager;
 };
