+++ TeamB_Shell/TeamB_Shell.vcxproj.filters
@@ -42,6 +42,21 @@
     <ClCompile Include="logger.cpp">
       <Filter>소스 파일</Filter>
     </ClCompile>
+    <ClCompile Include="test_suite\Test_EraseAndWriteAging_4.cpp">
+      <Filter>소스 파일</Filter>
+    </ClCompile>
+    <ClCompile Include="test_suite\Test_FullWriteAndReadCompare_1.cpp">
+      <Filter>소스 파일</Filter>
+    </ClCompile>
+    <ClCompile Include="test_suite\Test_FullWriteFullReadFlush_5.cpp">
+      <Filter>소스 파일</Filter>
+    </ClCompile>
+    <ClCompile Include="test_suite\Test_PartialLBAWrite_2.cpp">
+      <Filter>소스 파일</Filter>
+    </ClCompile>
+    <ClCompile Include="test_suite\Test_WriteReadAging_3.cpp">
+      <Filter>소스 파일</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <None Include="packages.config" />
@@ -63,6 +78,9 @@
       <Filter>헤더 파일</Filter>
     </ClInclude>
     <ClInclude Include="logger.h">
+      <Filter>헤더 파일</Filter>
+    </ClInclude>
+    <ClInclude Include="RegisterHelper.h">
       <Filter>헤더 파일</Filter>
     </ClInclude>
   </ItemGroup>
