+++ TeamB_Shell/TestManager.cpp
@@ -13,6 +13,15 @@ using ::testing::Return;
 using ::testing::Sequence;
 using ::testing::Test;
 
+TestManager& TestManager::instance() {
+  static TestManager instance;
+  return instance;
+}
+
+void TestManager::clearTests() {
+    testCases.clear();
+}
+
 void TestManager::registerTest(const string& name, TestFn func) {
   testCases[name] = func;
 }
@@ -184,22 +193,23 @@ bool TestMock_FullWriteFullReadFlush_5() {
 class TestScriptFixture : public Test {
  public:
   void registerAllTestcases() {
-    testManager.registerTest("1_FullWriteAndReadCompare_Mock",
+    TestManager::instance().clearTests();
+    TestManager::instance().registerTest("1_FullWriteAndReadCompare_Mock",
                              TestMock_FullWriteAndReadCompare_1);
-    testManager.registerTest("2_PartialLBAWrite_Mock",
+    TestManager::instance().registerTest("2_PartialLBAWrite_Mock",
                              TestMock_PartialLBAWrite_2);
-    testManager.registerTest("3_WriteReadAging_Mock",
+    TestManager::instance().registerTest("3_WriteReadAging_Mock",
                              TestMock_WriteReadAging_3);
-    testManager.registerTest("4_EraseAndWriteAging_Mock",
+    TestManager::instance().registerTest("4_EraseAndWriteAging_Mock",
                              TestMock_EraseAndWriteAging_4);
-    testManager.registerTest("5_FullWriteFullReadFlush_Mock",
+    TestManager::instance().registerTest("5_FullWriteFullReadFlush_Mock",
                              TestMock_FullWriteFullReadFlush_5);
   }
 
   void runTestcase(const std::string& testName) {
-    int result = testManager.runTest(testName);
+    int result = TestManager::instance().runTest(testName);
     std::cout << "[Test: " << testName << "] " << result << std::endl;
-    EXPECT_EQ(result, testManager.PASS);
+    EXPECT_EQ(result, TestManager::instance().PASS);
   }
 
   void runPrefixTestcase(const std::string& testName) {
@@ -219,27 +229,24 @@ class TestScriptFixture : public Test {
                 << "] Invalid number format: " << e.what() << std::endl;
       return;
     }
-    int result = testManager.runTest(strNum + "_");
+    int result = TestManager::instance().runTest(strNum + "_");
     std::cout << "[Test: " << testName << "] " << result << std::endl;
-    EXPECT_EQ(result, testManager.PASS);
+    EXPECT_EQ(result, TestManager::instance().PASS);
   }
 
   void runAllTestcases() {
-    vector<string> tests = testManager.listTests();
+    vector<string> tests = TestManager::instance().listTests();
     for (const auto& test : tests) {
       runTestcase(test);
     }
   }
 
   void runAllPrefixTestcases() {
-    vector<string> tests = testManager.listTests();
+    vector<string> tests = TestManager::instance().listTests();
     for (const auto& test : tests) {
       runPrefixTestcase(test);
     }
   }
-
- private:
-  TestManager testManager;
 };
 
 // TestScripts
@@ -251,114 +258,4 @@ TEST_F(TestScriptFixture, RunTestAll) {
 TEST_F(TestScriptFixture, PrefixTest) {
   registerAllTestcases();
   runAllPrefixTestcases();
-}
-
-bool Test_FullWriteAndReadCompare_1() {
-  ShellTest realShell;
-  UserCommandQueue queue(realShell);
-
-  const int minLba = 0;
-  const int maxLba = 99;
-  const unsigned int groupSize = 5;
-
-  uint32_t seed = static_cast<uint32_t>(time(0));
-  for (int base = minLba; base <= maxLba; base += groupSize + 1) {
-    unsigned int pattern = patternGenerator(seed);
-
-    for (unsigned int offset = 0; offset < groupSize && base + offset <= maxLba;
-         ++offset) {
-      if (!queue.enqueueWrite(base + offset, pattern)) return false;
-    }
-    for (unsigned int offset = 0; offset < groupSize && base + offset <= maxLba;
-         ++offset) {
-      if (!queue.enqueueRead(base + offset, pattern)) return false;
-    }
-
-    if (!queue.flush()) return false;
-  }
-  return true;
-}
-
-bool Test_PartialLBAWrite_2() {
-  ShellTest realShell;
-  UserCommandQueue queue(realShell);
-
-  const std::vector<int> lbaOrder = {4, 0, 3, 1, 2};
-  const int loopCount = 30;
-  uint32_t seed = static_cast<uint32_t>(time(0));
-  for (int loop = 0; loop < loopCount; ++loop) {
-    unsigned int pattern = patternGenerator(seed);
-
-    // Write sequence
-    for (int lba : lbaOrder) {
-      if (!queue.enqueueWrite(lba, pattern)) return false;
-    }
-
-    // ReadCompare sequence
-    for (int lba : lbaOrder) {
-      if (!queue.enqueueRead(lba, pattern)) return false;
-    }
-  }
-
-  return queue.flush();
-}
-
-bool Test_WriteReadAging_3() {
-  ShellTest realShell;
-  UserCommandQueue queue(realShell);
-
-  const int maxLoopCount = 200;
-  unsigned int pat1, pat2;
-
-  uint32_t seed = static_cast<uint32_t>(time(0));
-  for (int loop = 0; loop < maxLoopCount; loop++) {
-    pat1 = patternGenerator(seed);
-    pat2 = patternGenerator(seed);
-    if (!queue.enqueueWrite(0, pat1)) return false;
-    if (!queue.enqueueWrite(99, pat2)) return false;
-
-    if (!queue.enqueueRead(0, pat1)) return false;
-    if (!queue.enqueueRead(99, pat2)) return false;
-  }
-
-  return queue.flush();
-}
-
-bool Test_EraseAndWriteAging_4() {
-  ShellTest realShell;
-  UserCommandQueue queue(realShell);
-
-  const int maxLoopCount = 30;
-  const int len = 2;
-  const int maxLba = 99;
-  unsigned int pat;
-
-  uint32_t seed = static_cast<uint32_t>(time(0));
-  for (int loop = 0; loop < maxLoopCount; loop++) {
-    if (!queue.enqueueEraseRange(0, len)) return false;
-    for (int startLba = len; startLba + len <= maxLba; startLba += len) {
-      pat = patternGenerator(seed);
-      if (!queue.enqueueWrite(startLba, pat)) return false;
-
-      pat = patternGenerator(seed);
-      if (!queue.enqueueWrite(startLba, pat)) return false;
-      if (!queue.enqueueEraseRange(startLba, startLba + len)) return false;
-    }
-  }
-
-  return queue.flush();
-}
-
-bool Test_FullWriteFullReadFlush_5() {
-  ShellTest realShell;
-  UserCommandQueue queue(realShell);
-  unsigned int pat;
-
-  uint32_t seed = static_cast<uint32_t>(time(0));
-  pat = patternGenerator(seed);
-  if (!queue.enqueueFullWrite(pat)) return false;
-  if (!queue.enqueueFullRead()) return false;
-  if (!queue.enqueueFlush()) return false;
-
-  return queue.flush();
 }
