+++ TeamB_Shell/TestManager.h
@@ -15,7 +15,8 @@ typedef bool (*TestFn)();
 
 class TestManager {
  public:
-  TestManager() = default;
+  static TestManager& instance();
+  void clearTests();
   void registerTest(const string& name, TestFn func);
   int runTest(const string& name);
   vector<string> listTests();
@@ -40,9 +41,3 @@ bool TestMock_PartialLBAWrite_2();
 bool TestMock_WriteReadAging_3();
 bool TestMock_EraseAndWriteAging_4();
 bool TestMock_FullWriteFullReadFlush_5();
-
-bool Test_FullWriteAndReadCompare_1();
-bool Test_PartialLBAWrite_2();
-bool Test_WriteReadAging_3();
-bool Test_EraseAndWriteAging_4();
-bool Test_FullWriteFullReadFlush_5();
