+++ TeamB_Shell/main.cpp
@@ -7,15 +7,7 @@ int main() {
   return RUN_ALL_TESTS();
 #endif
 
-  TestManager testManager;
-  testManager.registerTest("1_FullWriteAndReadCompare",
-                           +Test_FullWriteAndReadCompare_1);
-  testManager.registerTest("2_PartialLBAWrite", Test_PartialLBAWrite_2);
-  testManager.registerTest("3_WriteReadAging", Test_WriteReadAging_3);
-  testManager.registerTest("4_EraseAndWriteAging", Test_EraseAndWriteAging_4);
-  testManager.registerTest("5_FullWriteFullReadFlush", Test_FullWriteFullReadFlush_5);
-
-  ShellTest shellTest{&testManager};
+  ShellTest shellTest{};
   while (true) {
     std::string input;
     std::cout << "> ";
